Refactor: Organize assets and layout folders, update Header design
