+++ sui_client/src/Pages/Challenge_1.tsx
@@ -10,7 +10,7 @@ import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 import { tomorrow } from "react-syntax-highlighter/dist/esm/styles/prism";
 import { useAtomValue } from "jotai";
 import { packageIdAtom } from "../atom";
-import Header from "../components/Header";
+import Header from "../components/layout/Header";
 import ChallengeDescription from "../components/ChallengeDescription";
 import PurpleButton from "../components/PurpleButton"
 import RedButton from "../components/RedButton";
