+++ sui_client/src/components/PurpleButton.tsx
@@ -11,14 +11,14 @@ const PurpleButton: React.FC<PurpleButtonProps> = ({ onClick, children, text }) 
     <button
       onClick={onClick}
       className="bg-purple-600 
-                 text-white 
-                 px-6 py-3 
-                 rounded-lg 
-                 font-semibold 
-                 shadow-md 
-                 border-2 border-white 
-                 transition-transform 
-                 transform hover:scale-105 hover:shadow-lg"
+                  text-white 
+                  px-6 py-3 
+                  rounded-lg 
+                  font-semibold 
+                  shadow-md 
+                  border-2 border-white 
+                  transition-transform 
+                  transform hover:scale-105 hover:shadow-lg"  
     >
       {text || children}
     </button>
