--- sui_client/src/styles/buttonStyles.ts
@@ -1,6 +1,0 @@
-export const buttonBaseStyle =
-  "font-inter bg-indigo-500 text-white px-4 py-6 w-36 h-36 rounded-md font-semibold text-sm text-center transition-all duration-300 ease-in-out shadow-md flex flex-col items-center justify-center cursor-pointer";
-
-export const buttonHoverStyle = "transform translate-y-[-2px] shadow-lg";
-
-export const buttonDefaultStyle = "transform translate-y-0 shadow-md";
